✨ Feat: social login buttons UI
