+++ tailwind.config.ts
@@ -5,7 +5,11 @@ const rem0_100 = { ...Array.from(Array(101)).map((_, i) => `${i / 10}rem`) };
 const rem0_1200 = { ...Array.from(Array(1201)).map((_, i) => `${i / 10}rem`) };
 
 module.exports = {
-  content: ['./src/pages/**/*.{js,ts,jsx,tsx,mdx}', './src/components/**/*.{js,ts,jsx,tsx,mdx}', './src/app/**/*.{js,ts,jsx,tsx,mdx}'],
+  content: [
+    './src/pages/**/*.{js,ts,jsx,tsx,mdx}',
+    './src/components/**/*.{js,ts,jsx,tsx,mdx}',
+    './src/app/**/*.{js,ts,jsx,tsx,mdx}',
+  ],
   theme: {
     extend: {
       borderWidth: rem0_10,
@@ -38,6 +42,7 @@ module.exports = {
         green: {
           DEFAULT: '#86D549',
           bg: '#E7F7DB',
+          naver: '#03CF5D',
         },
         purple: '#760DDE',
         orange: {
@@ -55,6 +60,7 @@ module.exports = {
         },
         white: '#FFF',
         'beige-f9': '#F9F7F7',
+        yellow: '#FEE500',
       },
       screens: {
         tablet: { max: '1199px' },
